add test for localizedDiscount
